Moved button into AddLearnerModal
